pyproject.toml
[tool.poetry]
name = "bills-collector"
version = "0.1.0"
description = ""
authors = ["Neeraj Verma <github@smurfpandey.me>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
environs = "^10.3.0"
flask = "^3.0.2"
flask-login = "^0.6.3"
flask-migrate = "^4.0.5"
flask-sqlalchemy = "^3.1.1"
authlib = "^1.3.0"
requests = "^2.31.0"
flask-bcrypt = "^1.0.1"
sentry-sdk = {extras = ["flask"], version = "^2.0.1"}
celery = "^5.3.6"
psycopg = {extras = ["binary", "pool"], version = "^3.1.18"}
pydash = "^7.0.7"
pikepdf = "^8.13.0"
google-api-python-client = "^2.127.0"
google-auth = "^2.29.0"
flask-wtf = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
